[refactor🔨] : VerifyService 추상화
